Move `ContainerTypes` and `ComponentTypes` into `models/`
