+++ src/components/app/DefinitionPane/DefinitionPane.mobile.tsx
@@ -3,7 +3,7 @@ import styled, { StyledFunction } from 'styled-components';
 
 import { resolveNames, withProps } from '@src/styles/styleUtils';
 import { styles, Colors } from '@styles/index';
-import { DefinitionPane } from '@src/components/ComponentTypes';
+import { DefinitionPane } from '@src/models/ComponentTypes';
 import Upper from './Upper/Upper';
 import Lower from './Lower/Lower';
 import Locale from '@src/components/app/Locale/Locale';
